refactor(robo): merged `dotenv.ts` into `env.ts`
Load functions now also return the newly loaded object. We'll be using this shortly.
